Start virtual memory management.

+++ src/include/kernel/types.h
@@ -3,7 +3,8 @@
 
 #include <stdint.h>
 
-typedef uint64_t 	tid_t;
+typedef uint16_t 	tid_t;	// This is 16-bit for the VMM mapping of 
+				// stacks.  See VmmManager.
 struct task_t;
 
 typedef uint64_t	cpuid_t;
